correçao de bug
